+++ src/pymedia/commands/animated/cmd.py
@@ -2,7 +2,7 @@
 
 from pymedia.commands.animated.parameters import AnimatedParameters
 from pymedia.errors import MissingParameterError
-from pymedia.types import OverwriteMode
+from pymedia.types import OverwriteMode, ScaleFlag
 
 
 class AnimatedCmd:
@@ -47,6 +47,7 @@ class AnimatedCmd:
                 str(self.params.media.path),
                 "-filter_complex",
                 self._build_filters(),
+                *self._build_args_per_animated_container(),
                 "-progress",
                 "pipe:1",
                 "-nostats",
@@ -58,21 +59,65 @@ class AnimatedCmd:
 
     def _build_filters(self) -> str:
         """Construye los filtros de ffmpeg para generar una imagen animada."""
-        output = self.params.animated_output
-        if output is None:
+        if self.params.animated_output is None:
             raise MissingParameterError(name="animated_output")
 
         filters: list[str] = []
 
-        filters_cmd = self.params.to_filters_cmd()
+        filters_cmd = self.params.to_filters_cmd(scale_flag=ScaleFlag.LANCZOS)
         if filters_cmd:
             filters.append(filters_cmd)
 
         filters.append(self.params.to_fps_cmd())
 
-        if output.suffix == ".gif":
-            filters.append(
-                "split[a][b];[a]palettegen[p];[b][p]paletteuse=dither=floyd_steinberg"
-            )
+        match self.params.animated_output.suffix:
+            case ".apng":
+                filters.append(
+                    "split[a][b];[a]palettegen=max_colors=256:stats_mode=diff[p];"
+                    "[b][p]paletteuse=dither=sierra2_4a"
+                )
+            case ".gif":
+                filters.append(
+                    "split[a][b];[a]palettegen[p];"
+                    "[b][p]paletteuse=dither=floyd_steinberg"
+                )
 
         return ",".join(filters)
+
+    def _build_args_per_animated_container(self) -> list[str]:
+        """Establece opciones dependiendo del contenedor de salida."""
+        if self.params.animated_output is None:
+            raise MissingParameterError(name="animated_output")
+
+        args: list[str] = []
+
+        match self.params.animated_output.suffix:
+            case ".apng":
+                return [
+                    "-c:v",
+                    "apng",
+                    "-compression_level",
+                    "9",
+                    "-plays",
+                    "0",
+                ]
+            case ".gif":
+                return [
+                    "-loop",
+                    "0",
+                ]
+            case ".webp":
+                return [
+                    "-c:v",
+                    "libwebp_anim",
+                    "-lossless",
+                    "0",
+                    "-quality",
+                    "80",
+                    "-compression_level",
+                    "6",
+                    "-loop",
+                    "0",
+                ]
+
+        return args